Refactor imports to use PortfolioPageContainer from the correct path and create PortfolioPageContainer component for better structure and organization.
