+++ src/app/portfolio/page.tsx
@@ -1,7 +1,7 @@
 'use client';
 import SidebarTriggerWithBreadcrumb from '@/components/SidebarTriggerWithBreadcrumb';
 import withAssetItems from '@/features/assetItems/hoc/withAssetItems';
-import { PortfolioPageContainer } from '@/features/portfolio/hoc/withPortfolios';
+import PortfolioPageContainer from '@/features/portfolio/components/PortfolioPageContainer';
 import InvestmentsFilterForm from '@/features/assetItems/components/InvestmentsFilterForm';
 import PortfolioOverallSection from '@/features/portfolio/components/PortfolioOverallSection';
 import PortfolioPerAssetItemSection from '@/features/portfolio/components/PortfolioPerAssetItemSection';
